feat(vesting): add vesting start and end dates
- remove unused components
